refactor(core): move computeEstimatedBudget/computeTripTotals into @btp/core (#1048)
